# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/providers": {
      "get": {
        "operationId": "AdminProviders_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List all AI providers",
        "tags": [
          "Admin: AI Providers"
        ]
      },
      "post": {
        "description": "Registers an OpenAI, Anthropic (Claude) or Google Gemini provider. The API key is encrypted at rest and never returned.",
        "operationId": "AdminProviders_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      }
    },
...
    "/api/v1/admin/providers/{id}": {
      "get": {
        "operationId": "AdminProviders_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      },
      "patch": {
        "description": "Send `apiKey` to rotate the key. The provider type cannot be changed.",
        "operationId": "AdminProviders_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Edit an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      },
      "delete": {
        "description": "Existing conversations keep their history; their provider reference is cleared.",
        "operationId": "AdminProviders_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Provider deleted"
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      }
    },
    "/api/v1/admin/providers/{id}/enable": {
      "post": {
        "operationId": "AdminProviders_enable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Enable an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      }
    },
    "/api/v1/admin/providers/{id}/disable": {
      "post": {
        "description": "Disabling the default provider also clears the default selection.",
        "operationId": "AdminProviders_disable",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Disable an AI provider",
        "tags": [
          "Admin: AI Providers"
        ]
      }
    },
    "/api/v1/admin/providers/{id}/default": {
      "post": {
        "operationId": "AdminProviders_makeDefault",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not allowed to perform this action",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflicts with existing state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Make this the default provider",
        "tags": [
          "Admin: AI Providers"
        ]
      }
    },
...
  },
  "info": {
    "title": "EchoGPT API",
    "description": "\nREST API powering the **EchoGPT** multi-AI chat Chrome extension.\n\n### Authentication\n1. `POST /api/v1/auth/register` or `POST /api/v1/auth/login` returns an **access token** (short lived JWT) and a **refresh token** (opaque, long lived, rotated on every use).\n2. Send the access token as `Authorization: Bearer <token>`. Click **Authorize** above to use it here.\n3. When the access token expires call `POST /api/v1/auth/refresh`. Re-using an already rotated refresh token revokes the whole session.\n\n### Errors\nEvery error uses the same envelope: `{ statusCode, error, message, details?, path, timestamp, requestId }`.\nThe `x-request-id` response header echoes the correlation id for support/debugging.\n\n### Quotas\nAI endpoints (chat, AI-assisted search) consume the caller's daily plan quota. Exceeding it returns **429**.\n",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Validation failed"
          },
          "details": {
            "example": [
              "email must be an email",
              "password must be longer than or equal to 8 characters"
            ],
            "description": "Field level details, present for validation errors.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/register"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-29T10:15:30.000Z"
          },
          "requestId": {
            "type": "string",
            "example": "6f1c2c1e-6c47-4f0e-9a36-0c5f0f4b1a2d"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp",
          "requestId"
        ]
      },
...
      "ProviderHealthDto": {
        "type": "object",
        "properties": {
          "status": {
            "enum": [
              "UNKNOWN",
              "HEALTHY",
              "UNHEALTHY"
            ],
            "type": "string",
            "example": "HEALTHY"
          },
          "checkedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "latencyMs": {
            "type": "number",
            "nullable": true,
            "example": 245
          },
          "error": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "status",
          "checkedAt",
          "latencyMs",
          "error"
        ]
      },
      "ProviderResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "OpenAI"
          },
          "type": {
            "enum": [
              "OPENAI",
              "ANTHROPIC",
              "GEMINI"
            ],
            "type": "string"
          },
          "baseUrl": {
            "type": "string",
            "example": "https://api.openai.com/v1"
          },
          "apiKeyMasked": {
            "type": "string",
            "example": "••••••••3xYz"
          },
          "defaultModel": {
            "type": "string",
            "example": "gpt-5-mini"
          },
          "models": {
            "example": [
              "gpt-5",
              "gpt-5-mini"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isEnabled": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "health": {
            "$ref": "#/components/schemas/ProviderHealthDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "baseUrl",
          "apiKeyMasked",
          "defaultModel",
          "models",
          "isEnabled",
          "isDefault",
          "health",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "OpenAI",
            "description": "Unique display name"
          },
          "type": {
            "enum": [
              "OPENAI",
              "ANTHROPIC",
              "GEMINI"
            ],
            "type": "string",
            "example": "OPENAI"
          },
          "apiKey": {
            "type": "string",
            "minLength": 8,
            "maxLength": 512,
            "description": "Vendor API key. Stored encrypted (AES-256-GCM) and never returned.",
            "example": "sk-proj-xxxxxxxxxxxxxxxxxxxxxxxx",
            "writeOnly": true
          },
          "baseUrl": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "description": "Override the vendor base URL (proxy / compatible gateway). Defaults per type.",
            "example": "https://api.openai.com/v1"
          },
          "models": {
            "minItems": 1,
            "maxItems": 50,
            "description": "Models users may select. Defaults to the vendor catalog.",
            "example": [
              "gpt-5",
              "gpt-5-mini"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            }
          },
          "defaultModel": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Must be one of `models`.",
            "example": "gpt-5-mini"
          },
          "isEnabled": {
            "type": "boolean",
            "default": true
          },
          "isDefault": {
            "type": "boolean",
            "default": false,
            "description": "Make this the default provider."
          }
        },
        "required": [
          "name",
          "type",
          "apiKey"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "example": "OpenAI",
            "description": "Unique display name"
          },
          "apiKey": {
            "type": "string",
            "minLength": 8,
            "maxLength": 512,
            "description": "Vendor API key. Stored encrypted (AES-256-GCM) and never returned.",
            "example": "sk-proj-xxxxxxxxxxxxxxxxxxxxxxxx",
            "writeOnly": true
          },
          "baseUrl": {
            "type": "string",
            "maxLength": 2048,
            "format": "uri",
            "description": "Override the vendor base URL (proxy / compatible gateway). Defaults per type.",
            "example": "https://api.openai.com/v1"
          },
          "models": {
            "minItems": 1,
            "maxItems": 50,
            "description": "Models users may select. Defaults to the vendor catalog.",
            "example": [
              "gpt-5",
              "gpt-5-mini"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            }
          },
          "defaultModel": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Must be one of `models`.",
            "example": "gpt-5-mini"
          }
        }
      },
...
    }
  }
}